requirements.txt
apache-airflow
apache-airflow-providers-docker
apache-airflow-providers-postgres
pandas
psycopg2-binary